+++ packages/paste-core/components/side-panel/src/SidePanelPushContentWrapper.tsx
@@ -50,11 +50,11 @@ export const SidePanelPushContentWrapper = React.forwardRef<HTMLDivElement, Side
       <AnimatedStyledContentWrapper
         {...safelySpreadBoxProps(props)}
         ref={ref}
-        element={element}
         // when using side panels in responsive layouts, we don't want any left margin in small screen, or initial SSR render situations. So basically never apply it in those situations
         style={breakpointIndex === (undefined || 0) ? undefined : styles}
         marginRight={["space0", theme.sizes.size40]}
         minWidth="size0"
+        element={element}
       >
         {children}
       </AnimatedStyledContentWrapper>
